requirements.txt
flask
flask-cors
boto3
python-dotenv
requests
werkzeug
